requirements.txt
certstream==1.10
PyYAML==5.1
python_Levenshtein==0.12.0
tqdm==4.19.4
tldextract~=3.1.0
requests~=2.25.1
python-dateutil~=2.8.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
numpy~=1.19.5
sklearn~=0.0
scikit-learn~=0.24.2
pandas~=0.25.3
matplotlib~=3.3.4
seaborn~=0.11.1